rename(route-folder): chatroom utils 배럴 패턴 제거
