+++ src/api/fetch/chatMessage/api/useSendImage.ts
@@ -1,7 +1,10 @@
 import useAppMutation from "@/api/_base/query/useAppMutation";
 import { useQueryClient } from "@tanstack/react-query";
 import { ChatMessage } from "../types/ChatMessageResponse";
-import { addMessageToCache, removeMessageFromCache } from "@/app/(route)/chat/[postId]/_utils";
+import {
+  addMessageToCache,
+  removeMessageFromCache,
+} from "@/app/(route)/chat/[postId]/_utils/chatMessageCache/chatMessageCache";
 
 interface SendImageContext {
   optimisticId: number;
