+++ src/app/(route)/chat/[postId]/_hooks/useChatMessageSubmit/useChatMessageSubmit.test.tsx
@@ -1,8 +1,11 @@
 import { act, renderHook } from "@testing-library/react";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { sendChatSocketMessage } from "@/api/fetch/chatRoom";
-import { addMessageToCache, removeMessageFromCache } from "../../_utils";
 import useChatMessageSubmit from "./useChatMessageSubmit";
+import {
+  addMessageToCache,
+  removeMessageFromCache,
+} from "../../_utils/chatMessageCache/chatMessageCache";
 
 jest.mock("@/api/fetch/chatRoom", () => ({
   sendChatSocketMessage: jest.fn(),
