+++ src/app/(route)/chat/[postId]/_hooks/useChatMessageSubmit/useChatMessageSubmit.ts
@@ -1,7 +1,10 @@
 import { UseFormReset } from "react-hook-form";
 import { sendChatSocketMessage } from "@/api/fetch/chatRoom";
 import { ChatMessage } from "@/api/fetch/chatMessage/types/ChatMessageResponse";
-import { addMessageToCache, removeMessageFromCache } from "../../_utils";
+import {
+  addMessageToCache,
+  removeMessageFromCache,
+} from "../../_utils/chatMessageCache/chatMessageCache";
 import { useQueryClient } from "@tanstack/react-query";
 
 interface UseChatMessageSubmitParams {
